Yoni - added python and gradle
