--- src/main/java/frc/utils/applicationsutils/test.java
@@ -1,13 +1,0 @@
-package frc.utils.applicationsutils;
-
-import javax.swing.*;
-
-public class test {
-
-    public static void main(String[] args) {
-        JFrame frame = new JFrame();
-        frame.setVisible(true);
-        frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-        frame.setSize(500,500);
-    }
-}
